+++ src/compas/datastructures/mesh/join.py
@@ -41,7 +41,7 @@ def mesh_weld(mesh, precision=None, cls=None):
     gkey_index = {gkey: index for index, gkey in enumerate(gkey_key)}
     vertices = [key_xyz[key] for gkey, key in gkey_key.items()]
     faces = [[gkey_index[geo(key_xyz[key], precision)] for key in mesh.face_vertices(fkey)] for fkey in mesh.faces()]
-    faces = [[u for u, v in pairwise(face + face[:1]) if u != v] for face in faces]
+    faces[:] = [[u for u, v in pairwise(face + face[:1]) if u != v] for face in faces]
 
     return cls.from_vertices_and_faces(vertices, faces)
 
@@ -115,6 +115,17 @@ def meshes_join_and_weld(meshes, precision=None, cls=None):
 
 if __name__ == "__main__":
 
-    import compas
+    from compas.datastructures import Mesh
+    from compas.datastructures import mesh_weld
+    from compas_plotters import MeshPlotter
 
+    vertices = [[0, 0, 0], [0.04, 0, 0], [1.0, 0, 0], [1.0, 1.0, 0], [0, 1.0, 0]]
+    faces = [[0, 1, 2, 3, 4]]
 
+    mesh = Mesh.from_vertices_and_faces(vertices, faces)
+    mesh = mesh_weld(mesh, precision='1f')
+
+    plotter = MeshPlotter(mesh, figsize=(10, 7))
+    plotter.draw_vertices(text={key: "{:.3f}".format(mesh.vertex[key]['x']) for key in mesh.vertices()}, radius=0.03)
+    plotter.draw_edges()
+    plotter.show()